refactor(framework/ai-sdk): modularize history/reply mappers and clarify tests
Split history-mapper and reply-mapper into focused submodules, improve test
fixtures and scenario naming, and expand ai-sdk-agent adapter coverage.
